Updated custom widget for accounts listing, added unit test
