+++ lib/pages/home_page.dart
@@ -1,12 +1,11 @@
 // Home page.
 
 import 'dart:ui';
-import "dart:math";
 
 import 'package:flutter/material.dart';
 import 'package:percent_indicator/percent_indicator.dart';
 
-import 'package:sossoldi/customWidgets/accounts_sum.dart';
+import 'package:sossoldi/custom_widgets/accounts_sum.dart';
 
 class HomePage extends StatefulWidget {
   @override
@@ -96,7 +95,7 @@ class _HomePageState extends State<HomePage> {
                   "Your accounts",
                   style: Theme.of(context).textTheme.displayMedium,
                 ))),
-        const SizedBox(height: 10),
+        const SizedBox(height: 16),
         Container(
           height: 85.0,
           margin: const EdgeInsets.only(left: 16.0, right: 16.0),
@@ -105,10 +104,7 @@ class _HomePageState extends State<HomePage> {
               shrinkWrap: true,
               scrollDirection: Axis.horizontal,
               itemBuilder: (context, i) {
-                var accountsList = ['N26','Fineco','Crypto.com','Mediolanum'];
-                var amountsList = ['3.823,56','0,07','574,22','14.549,01'];
-                final random = Random();
-                return AccountsSum(accountName: accountsList[random.nextInt(accountsList.length)], amount: amountsList[random.nextInt(amountsList.length)]);
+                return const AccountsSum(accountName: 'Cash', amount: '1.234,56');
               }),
         ),
         const SizedBox(height: 28),
